Don't show lock screen before unlock screen
Currently, when the phone is in a locked state (for example, just
after a reboot) the lockscreen is too small because of the presence
of the emergy dailer. This causes a push down of the multiwave view,
making it not fully visible. If the lockscreen is in the future some-
how modified, an other object can make the multiwave view totally
unvisible, causing a lock down of the device as you can't get to the
"Enter PIN" screen.

We already have had this feature when the lockscreen wasn't resized
yet, and the weather widget was visible. This commit is somewhat
a bye-pass of that problem rather than a fix. Nevertheless, is does
the trick.

Change-Id: Id22e99d81084510177e2f6afde316c6f9e042fbb